+++ include/private/contingencylist.h
@@ -34,10 +34,18 @@ struct _p_Contingency {
 typedef struct _p_Contingency Contingency;
 
 struct _p_ContingencyList {
-  PetscInt Ncont; /* Number of contingencies = number of scenarios */
+  PetscInt    Ncontinit; /* Initial size of list */
+  PetscInt    Ncont; /* Number of contingencies = number of scenarios */
   Contingency *cont; /* Contingencies */
+  char        inputfile[PETSC_MAX_PATH_LEN]; /* input file */
+  ContingencyFileInputFormat inputfileformat; /* format of input file */
 };
 
-typedef struct _p_ContingencyList ContingencyList;
+typedef struct _p_ContingencyList *ContingencyList;
+
+extern PetscErrorCode ContingencyListSetData(ContingencyList,ContingencyFileInputFormat,const char[]);
+extern PetscErrorCode ContingencyListCreate(PetscInt,ContingencyList*);
+extern PetscErrorCode ContingencyListDestroy(ContingencyList*);
+extern PetscErrorCode ContingencyListReadData(ContingencyList,PetscInt*);
 
 #endif