+++ src/sopflow/interface/sopflowscen.c
@@ -2,6 +2,7 @@
 #include <private/tcopflowimpl.h>
 #include <private/scopflowimpl.h>
 #include <private/sopflowimpl.h>
+#include <private/scenariolist.h>
 
 extern void clean2Char(char *);
 extern char** blankTokenizer(const char *str, int *numtok, int maxtokens, int maxchar);
@@ -30,49 +31,142 @@ PetscErrorCode SOPFLOWSetScenarioData(SOPFLOW sopflow,ScenarioFileInputFormat sc
 }
 
 /*
-  SOPFLOWReadScenarioData_Wind - Reads the wind data 
+  SOPFLOWReadScenarioData_Wind_SinglePeriod - Reads the wind data and populates the scenario list
   Input Parameters
 + sopflow - SOPFLOW object
 . windgenprofile - wind generator profile file
-. c       - contingency number
-. t - time step
+
+  Note: This function reads the wind scenario data for "single period" format files.
 */
-PetscErrorCode SOPFLOWReadScenarioData_Wind(SOPFLOW sopflow,const char windgenprofile[],PetscInt c,PetscInt t)
+PetscErrorCode SOPFLOWReadScenarioData_Wind_SinglePeriod(SOPFLOW sopflow,const char windgenprofile[])
 {
   PetscErrorCode ierr;
   FILE           *fp;
   char           line[MAXLINE];
   char           *out;
-  OPFLOW         opflow;
-  PS             ps;
-  PetscInt       ngen,*windgenbus,nw=0;
+  PetscInt       ngenwind,nw=0;
   char           *tok,*tok2;
   char           sep[] = ",",sep2[] = "_";
-  PSGEN          gen;
-  PetscReal      pg;
+  PetscReal      pg,weight;
   int            scen_num=0;
   int            genid;
+  int            windgenbus[100];
   char           windgenid[100][3];
-  int            scensread=0;
-  int            nscens;
+  int            i;
+  ScenarioList   *scenlist=&sopflow->scenlist;
+  Scenario       *scenario;
+  Forecast       *forecast;
 
   PetscFunctionBegin;
 
-  if(!sopflow->ismulticontingency) ngen = sopflow->opflows[0]->ps->ngen;
-  else {
-    if(!sopflow->scopflows[0]->ismultiperiod) ngen = sopflow->scopflows[0]->opflows[c]->ps->ngen;
-    else ngen = sopflow->scopflows[0]->tcopflows[c]->opflows[t]->ps->ngen;
-  }
+  ngenwind = 100; // This should be increased for larger cases (?)
   fp = fopen(windgenprofile,"r");
   if (fp == NULL) {
     SETERRQ1(PETSC_COMM_SELF,PETSC_ERR_FILE_OPEN,"Cannot open wind generation profile file %s",windgenprofile);CHKERRQ(ierr);
   }
 
-  ierr = PetscMalloc1(ngen,&windgenbus);CHKERRQ(ierr);
+  /* First line -- has the bus numbers */
+  out = fgets(line,MAXLINE,fp);
+  /* Parse wind generator numbers */
+  tok = strtok(line,sep);
+  tok = strtok(NULL,sep); /* Skip first token */
+  while(tok != NULL) {
+    if(strcmp(tok,"weight") == 0 || strcmp(tok,"weight\n") == 0) {
+      tok = strtok(NULL,sep);
+      continue;
+    }
+    /* Parse generator info */
+    tok2 = strsep(&tok,sep2);
+    sscanf(tok2,"%d",&windgenbus[nw]);
+    tok2 = strsep(&tok,sep2);
+    tok2 = strsep(&tok,sep2); /* Skip string "Wind" */
+    sscanf(tok2,"%d",&genid);
+    if(nw == ngenwind) SETERRQ1(PETSC_COMM_SELF,PETSC_ERR_SUP,"Exceeded max. number of wind generators=%d\n",ngenwind);
+    snprintf(windgenid[nw],3,"%-2d",genid);
 
-  /* scenarios on this rank excluding base */
-  //  nscens=sopflow->comm->rank?sopflow->ns:sopflow->ns-1;
-  nscens = sopflow->ns;
+    nw++;
+    tok = strtok(NULL,sep);
+  }
+
+  while((out = fgets(line,MAXLINE,fp)) != NULL) {
+    if(strcmp(line,"\r\n") == 0 || strcmp(line,"\n") == 0) {
+      continue; /* Skip blank lines */
+    }
+
+    tok = strtok(line,sep);
+    sscanf(tok,"%d",&scen_num); /* Scenario number */
+    scen_num -= 1; /* Scenario numbers start with 1 in the file, convert to zero-based start */
+
+    if(scen_num < scenlist->Nscen || scen_num == sopflow->Ns) {
+      fclose(fp);
+      PetscFunctionReturn(0);
+    }
+
+    scenario = &scenlist->scen[scen_num];
+    forecast = &scenario->forecastlist[scenario->nforecast];
+    forecast->num = scen_num;
+    forecast->type = FORECAST_WIND;
+    forecast->nele = nw;
+    ierr = PetscCalloc1(forecast->nele,&forecast->buses);CHKERRQ(ierr);
+    ierr = PetscCalloc1(forecast->nele,&forecast->id);CHKERRQ(ierr);
+    for(i=0; i < forecast->nele; i++) {
+      ierr = PetscCalloc1(3,&forecast->id[i]);
+    }
+    ierr = PetscCalloc1(forecast->nele,&forecast->val);CHKERRQ(ierr);
+    
+    tok = strtok(NULL,sep);
+    for(i=0; i < nw; i++) {
+      forecast->buses[i] = windgenbus[i];
+      ierr = PetscStrcpy(forecast->id[i],windgenid[i]);CHKERRQ(ierr);
+      sscanf(tok,"%lf",&pg);
+      forecast->val[i] = pg;
+      tok = strtok(NULL,sep);
+    }
+
+    /* Read scenario weight */
+    sscanf(tok,"%lf",&weight);
+    scenario->prob = weight;
+
+    scenario->nforecast++;
+    scenlist->Nscen++;
+  }
+  PetscFunctionReturn(0);
+}
+
+/*
+  SOPFLOWReadScenarioData_Wind_MultiPeriod - Reads the wind data and populates the scenario list
+  Input Parameters
++ sopflow - SOPFLOW object
+. windgenprofile - wind generator profile file
+
+  Note: This function reads the wind scenario data for "multi-period" format files.
+*/
+PetscErrorCode SOPFLOWReadScenarioData_Wind_MultiPeriod(SOPFLOW sopflow,const char windgenprofile[])
+{
+  PetscErrorCode ierr;
+  FILE           *fp;
+  char           line[MAXLINE];
+  char           *out;
+  PetscInt       ngenwind,nw=0;
+  char           *tok,*tok2;
+  char           sep[] = ",",sep2[] = "_";
+  PetscReal      pg;
+  int            scen_num=0;
+  int            genid;
+  int            windgenbus[100];
+  char           windgenid[100][3];
+  int            i;
+  ScenarioList   *scenlist=&sopflow->scenlist;
+  Scenario       *scenario;
+  Forecast       *forecast;
+
+  PetscFunctionBegin;
+
+  ngenwind = 100; // This should be increased for larger cases (?)
+  fp = fopen(windgenprofile,"r");
+  if (fp == NULL) {
+    SETERRQ1(PETSC_COMM_SELF,PETSC_ERR_FILE_OPEN,"Cannot open wind generation profile file %s",windgenprofile);CHKERRQ(ierr);
+  }
 
   /* First line -- has the bus numbers */
   out = fgets(line,MAXLINE,fp);
@@ -87,7 +181,7 @@ PetscErrorCode SOPFLOWReadScenarioData_Wind(SOPFLOW sopflow,const char windgenpr
     tok2 = strsep(&tok,sep2);
     tok2 = strsep(&tok,sep2); /* Skip string "Wind" */
     sscanf(tok2,"%d",&genid);
-    if(nw == 100) SETERRQ(PETSC_COMM_SELF,PETSC_ERR_SUP,"Exceeded max. values set for parsing wind generators\n");
+    if(nw == ngenwind) SETERRQ1(PETSC_COMM_SELF,PETSC_ERR_SUP,"Exceeded max. number of wind generators=%d\n",ngenwind);
     snprintf(windgenid[nw],3,"%-2d",genid);
 
     nw++;
@@ -104,42 +198,42 @@ PetscErrorCode SOPFLOWReadScenarioData_Wind(SOPFLOW sopflow,const char windgenpr
     sscanf(tok,"%d",&scen_num); /* Scenario number */
     scen_num -= 1; /* Scenario numbers start with 1 in the file, convert to zero-based start */
 
-    if(scensread == nscens) {
-      ierr = PetscFree(windgenbus);CHKERRQ(ierr);
+    if(scen_num < scenlist->Nscen || scen_num == sopflow->Ns) {
       fclose(fp);
       PetscFunctionReturn(0);
     }
 
-    if(scen_num < sopflow->sstart) { /* Haven't yet reached scenarios for this rank */
-      continue;
-    } else if(sopflow->sstart <= scen_num && scen_num < sopflow->send) { /* This is my range */
-      scen_num -= sopflow->sstart; /* convert to local scenario number */
-      scensread++;
-    } else { /* Beyond range, we can exit */
-      ierr = PetscFree(windgenbus);CHKERRQ(ierr);
-      fclose(fp);
-      PetscFunctionReturn(0);
+    scenario = &scenlist->scen[scen_num];
+    forecast = &scenario->forecastlist[scenario->nforecast];
+    forecast->num = scen_num;
+    forecast->type = FORECAST_WIND;
+    forecast->nele = nw;
+    ierr = PetscCalloc1(forecast->nele,&forecast->buses);CHKERRQ(ierr);
+    ierr = PetscCalloc1(forecast->nele,&forecast->id);CHKERRQ(ierr);
+    for(i=0; i < forecast->nele; i++) {
+      ierr = PetscCalloc1(3,&forecast->id[i]);
+    }
+    ierr = PetscCalloc1(forecast->nele,&forecast->val);CHKERRQ(ierr);
+    
+    for(i=0; i < nw; i++) {
+      forecast->buses[i] = windgenbus[i];
+      ierr = PetscStrcpy(forecast->id[i],windgenid[i]);CHKERRQ(ierr);
     }
 
-    if(!sopflow->ismulticontingency) opflow = sopflow->opflows[scen_num];
-    else {
-      if(!sopflow->scopflows[scen_num]->ismultiperiod) opflow = sopflow->scopflows[scen_num]->opflows[c];
-      else opflow = sopflow->scopflows[scen_num]->tcopflows[c]->opflows[t];
-    }
-    ps     = opflow->ps;
     tok = strtok(NULL,sep);
     nw = 0;
     while(tok != NULL) {
-      ierr = PSGetGen(ps,windgenbus[nw],windgenid[nw],&gen);CHKERRQ(ierr);
       sscanf(tok,"%lf",&pg);
-      gen->pg = gen->pt = pg/ps->MVAbase; /* Set real power generation. Note that Pg upper limit is also set to Pg */
-      gen->pgs = gen->pb;
+      forecast->val[nw] = pg;
       nw++;
       tok = strtok(NULL,sep);
     }
+    scenario->nforecast++;
+    scenlist->Nscen++;
   }
   PetscFunctionReturn(0);
 }
+
   
 /*
   SOPFLOWReadScenarioData - Reads the scenario data file
@@ -153,35 +247,22 @@ PetscErrorCode SOPFLOWReadScenarioData_Wind(SOPFLOW sopflow,const char windgenpr
 PetscErrorCode SOPFLOWReadScenarioData(SOPFLOW sopflow,ScenarioFileInputFormat scenfileformat,const char scenfile[])
 {
   PetscErrorCode ierr;
-  SCOPFLOW       scopflow;
-  TCOPFLOW       tcopflow;
-  PetscInt       c,t;
 
   PetscFunctionBegin;
   if(sopflow->scenunctype == WIND) {
-    if(!sopflow->ismulticontingency) {
-      ierr = SOPFLOWReadScenarioData_Wind(sopflow,scenfile,0,0);CHKERRQ(ierr);
-    } else {
-      scopflow = sopflow->scopflows[0];
-      for(c=0; c < scopflow->nc; c++) {
-	if(!scopflow->ismultiperiod) {
-	  ierr = SOPFLOWReadScenarioData_Wind(sopflow,scenfile,c,0);CHKERRQ(ierr);
-	} else {
-	  tcopflow = scopflow->tcopflows[0];
-	  for(t=0; t < tcopflow->Nt; t++) {
-	    ierr = SOPFLOWReadScenarioData_Wind(sopflow,scenfile,c,t);CHKERRQ(ierr);
-	  }
-	}
-      }
-    } 
+    if(scenfileformat == SOPFLOW_NATIVE_SINGLEPERIOD) {
+      ierr = SOPFLOWReadScenarioData_Wind_SinglePeriod(sopflow,scenfile);CHKERRQ(ierr);
+    } else if(scenfileformat == SOPFLOW_NATIVE_MULTIPERIOD) {
+      ierr = SOPFLOWReadScenarioData_Wind_MultiPeriod(sopflow,scenfile);CHKERRQ(ierr);
+    }
   }
 
   PetscFunctionReturn(0);
 }
 
-/* SOPFLOWGetNumScenarios_Native - Gets the number of scenarios from the scenario file
+/* SOPFLOWGetNumScenarios_Native_MultiPeriod - Gets the number of scenarios from the scenario file
  */
-PetscErrorCode SOPFLOWGetNumScenarios_Native(const char scenfile[],PetscInt *Ns)
+PetscErrorCode SOPFLOWGetNumScenarios_Native_MultiPeriod(const char scenfile[],PetscInt *Ns)
 {
   PetscErrorCode ierr;
   FILE           *fp;
@@ -219,6 +300,42 @@ PetscErrorCode SOPFLOWGetNumScenarios_Native(const char scenfile[],PetscInt *Ns)
   PetscFunctionReturn(0);
 }
 
+/* SOPFLOWGetNumScenarios_Native_SinglePeriod - Gets the number of scenarios from the scenario file
+ */
+PetscErrorCode SOPFLOWGetNumScenarios_Native_SinglePeriod(const char scenfile[],PetscInt *Ns)
+{
+  PetscErrorCode ierr;
+  FILE           *fp;
+  char           line[MAXLINE];
+  char           *out;
+  char           *tok,*tok2;
+  char           sep[] = ",",sep2[] = "_";
+  int            Nscen=0;
+
+  PetscFunctionBegin;
+
+  fp = fopen(scenfile,"r");
+  if (fp == NULL) {
+    SETERRQ1(PETSC_COMM_SELF,PETSC_ERR_FILE_OPEN,"Cannot open wind generation profile file %s",scenfile);CHKERRQ(ierr);
+  }
+
+  /* First line -- has the bus numbers */
+  out = fgets(line,MAXLINE,fp);
+
+  while((out = fgets(line,MAXLINE,fp)) != NULL) {
+    if(strcmp(line,"\r\n") == 0 || strcmp(line,"\n") == 0) {
+      continue; /* Skip blank lines */
+    }
+
+    tok = strtok(line,sep);
+    sscanf(tok,"%d",&Nscen); /* Scenario number */
+  }
+
+  fclose(fp);
+  *Ns = Nscen;
+  PetscFunctionReturn(0);
+}
+
 /*
   SOPFLOWReadScenarioData - Gets the number of scenarios from the scenario data file
 
@@ -234,8 +351,10 @@ PetscErrorCode SOPFLOWGetNumScenarios(SOPFLOW sopflow,ScenarioFileInputFormat sc
   PetscErrorCode ierr;
 
   PetscFunctionBegin;
-  if(scenfileformat == SOPFLOW_NATIVE) {
-    ierr = SOPFLOWGetNumScenarios_Native(scenfile,Ns);CHKERRQ(ierr);
+  if(scenfileformat == SOPFLOW_NATIVE_SINGLEPERIOD) {
+    ierr = SOPFLOWGetNumScenarios_Native_SinglePeriod(scenfile,Ns);CHKERRQ(ierr);
+  } else if(scenfileformat == SOPFLOW_NATIVE_MULTIPERIOD) {
+    ierr = SOPFLOWGetNumScenarios_Native_MultiPeriod(scenfile,Ns);CHKERRQ(ierr);
   }
   else *Ns = 1;